+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Renamed `ConditionCov` termination criterion to `TolConditionCov` and added `tol_condition_cov` option to configure it.
 - Added `MaxTime` termination criterion to limit the running time of the algorithm.
 - Renamed `EqualFunValues` termination criterion to `TolFunHist` and added `tol_fun_hist` option to configure its range.
+- Added `TolFunRel` termination criterion to allow for a relative tolerance in function.
 
 ## 0.1.1 (March 4th, 2022)